16 bit different minesweeper boards, minesweeper can print mines left, minesweeper print time when finished

+++ da64/keymaps/default/da_util.h
@@ -22,7 +22,7 @@
 #define BUFFER_SIZE 16
 
 uint16_t timer;
-uint8_t random_counter;
+uint16_t random_counter;
 
 // call this
 void da_util_matrix_scan_user(void);
